Rename to createScrollParent
